+++ zcash_client_sqlite/src/wallet/init/migrations/add_transaction_views.rs
@@ -278,7 +278,7 @@ mod tests {
     use zip32::AccountId;
 
     use crate::{
-        wallet::init::{init_wallet_db_internal, migrations::addresses_table},
+        wallet::init::{migrations::addresses_table, WalletMigrator},
         WalletDb,
     };
 
@@ -303,7 +303,9 @@ mod tests {
         let network = Network::TestNetwork;
         let data_file = NamedTempFile::new().unwrap();
         let mut db_data = WalletDb::for_path(data_file.path(), network, ()).unwrap();
-        init_wallet_db_internal(&mut db_data, None, &[addresses_table::MIGRATION_ID], false)
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(&mut db_data, &[addresses_table::MIGRATION_ID])
             .unwrap();
         let usk = UnifiedSpendingKey::from_seed(&network, &[0u8; 32][..], AccountId::ZERO).unwrap();
         let ufvk = usk.to_unified_full_viewing_key();
@@ -335,7 +337,10 @@ mod tests {
             VALUES (0, 4, 0, '', 7, '', 'c', true, X'63');",
         ).unwrap();
 
-        init_wallet_db_internal(&mut db_data, None, &[super::MIGRATION_ID], false).unwrap();
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(&mut db_data, &[super::MIGRATION_ID])
+            .unwrap();
 
         let mut q = db_data
             .conn
@@ -402,13 +407,13 @@ mod tests {
         let network = Network::TestNetwork;
         let data_file = NamedTempFile::new().unwrap();
         let mut db_data = WalletDb::for_path(data_file.path(), network, ()).unwrap();
-        init_wallet_db_internal(
-            &mut db_data,
-            None,
-            &[utxos_table::MIGRATION_ID, ufvk_support::MIGRATION_ID],
-            false,
-        )
-        .unwrap();
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(
+                &mut db_data,
+                &[utxos_table::MIGRATION_ID, ufvk_support::MIGRATION_ID],
+            )
+            .unwrap();
 
         // create a UTXO to spend
         let tx = TransactionData::from_parts(
@@ -479,7 +484,10 @@ mod tests {
             )
             .unwrap();
 
-        init_wallet_db_internal(&mut db_data, None, &[super::MIGRATION_ID], false).unwrap();
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(&mut db_data, &[super::MIGRATION_ID])
+            .unwrap();
 
         let fee = db_data
             .conn